requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cron==0.5.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==0.21.0
whitenoise==5.3.0
django-environ==0.9.0
